requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
pytz==2024.1
gunicorn==21.2.0
google-api-python-client==2.104.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
google-auth==2.23.3
psycopg2-binary==2.9.10
pymysql==1.1.2
pyodbc==5.3.0
beautifulsoup4==4.12.2
lxml==5.3.0
